IGNITE-16498 / rename packages
