+++ modules/rest-api/src/main/java/org/apache/ignite/internal/rest/api/RestHandlersRegister.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.rest;
+package org.apache.ignite.internal.rest.api;
 
 import java.util.function.Consumer;
