+++ modules/rest-api/src/test/java/org/apache/ignite/internal/rest/api/RestApiHttpResponseTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.ignite.rest;
+package org.apache.ignite.internal.rest.api;
 
 import static io.netty.handler.codec.http.HttpHeaderNames.CONTENT_TYPE;
 import static io.netty.handler.codec.http.HttpHeaderValues.APPLICATION_JSON;
@@ -26,8 +26,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import io.netty.handler.codec.http.DefaultHttpResponse;
 import io.netty.handler.codec.http.HttpVersion;
 import java.util.Map;
-import org.apache.ignite.rest.ErrorResult;
-import org.apache.ignite.rest.RestApiHttpResponse;
 import org.junit.jupiter.api.Test;
 
 /**
