+++ modules/rest/src/test/java/org/apache/ignite/internal/rest/routes/RouteTest.java
@@ -28,8 +28,8 @@ import io.netty.handler.codec.http.DefaultHttpRequest;
 import io.netty.handler.codec.http.HttpHeaderNames;
 import io.netty.handler.codec.http.HttpHeaderValues;
 import java.util.concurrent.CompletableFuture;
-import org.apache.ignite.rest.RequestHandler;
-import org.apache.ignite.rest.Route;
+import org.apache.ignite.internal.rest.api.RequestHandler;
+import org.apache.ignite.internal.rest.api.Route;
 import org.junit.jupiter.api.Test;
 
 /**
